+++ solstice/CHANGELOG.md
@@ -5,6 +5,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 ## [Unreleased]
 ### Fixed
 - `Solstice.start(String)` now throws an exception if there was no bundle with that name.
+- Locations registered in `ShimIdeBootstrapServices` have `url` property to keep Eclipse 4.25+ happy.
 
 ## [0.19.1] - 2023-03-12
 ### Fixed