+++ solstice/CHANGELOG.md
@@ -3,6 +3,8 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
+
+## [0.19.2] - 2023-03-13
 ### Fixed
 - `Solstice.start(String)` now throws an exception if there was no bundle with that name. ([#107](https://github.com/equodev/equo-ide/pull/107))
 - Locations registered in `ShimIdeBootstrapServices` have `url` property to keep Eclipse 4.25+ happy. ([#107](https://github.com/equodev/equo-ide/pull/107))